Move dependency merging from Project to Manifest.
